+++ src/promptflow-evals/tests/evals/e2etests/test_adv_simulator.py
@@ -223,7 +223,7 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=True,
+                _jailbreak_type="upia",
             )
         )
         assert len(outputs) == 1
@@ -265,7 +265,7 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=True,
+                _jailbreak_type="upia",
             )
         )
         assert len(outputs) == 1
@@ -310,7 +310,6 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=False,
             )
         )
         assert len(outputs) == 1
@@ -356,7 +355,46 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=False,
+            )
+        )
+        assert len(outputs) == 1
+
+    @pytest.mark.skipif(
+        not is_replay(), reason="API not fully released yet. Don't run in live mode unless connected to INT."
+    )
+    @pytest.mark.usefixtures("vcr_recording")
+    def test_adv_xpia_sim_responds_with_responses(self, azure_cred, project_scope):
+        os.environ.pop("RAI_SVC_URL", None)
+        from promptflow.evals.synthetic import AdversarialScenario, IndirectAttackSimulator
+
+        azure_ai_project = {
+            "subscription_id": project_scope["subscription_id"],
+            "resource_group_name": project_scope["resource_group_name"],
+            "project_name": project_scope["project_name"],
+        }
+
+        async def callback(
+            messages: List[Dict], stream: bool = False, session_state: Any = None, context: Dict[str, Any] = None
+        ) -> dict:
+            question = messages["messages"][0]["content"]
+
+            formatted_response = {"content": question, "role": "assistant"}
+            messages["messages"].append(formatted_response)
+            return {
+                "messages": messages["messages"],
+                "stream": stream,
+                "session_state": session_state,
+                "context": context,
+            }
+
+        simulator = IndirectAttackSimulator(azure_ai_project=azure_ai_project, credential=azure_cred)
+
+        outputs = asyncio.run(
+            simulator(
+                scenario=AdversarialScenario.ADVERSARIAL_INDIRECT_JAILBREAK,
+                max_conversation_turns=1,
+                max_simulation_results=1,
+                target=callback,
             )
         )
         assert len(outputs) == 1
@@ -398,7 +436,7 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=True,
+                _jailbreak_type="upia",
                 randomization_seed=1,
             )
         )
@@ -413,7 +451,7 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=True,
+                _jailbreak_type="upia",
                 randomization_seed=1,
             )
         )
@@ -428,7 +466,7 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=True,
+                _jailbreak_type="upia",
                 randomization_seed=2,
             )
         )
@@ -473,7 +511,6 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=False,
                 randomization_seed=1,
             )
         )
@@ -488,7 +525,6 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=False,
                 randomization_seed=1,
             )
         )
@@ -503,7 +539,6 @@ class TestAdvSimulator:
                 api_call_retry_sleep_sec=1,
                 api_call_delay_sec=30,
                 concurrent_async_task=1,
-                jailbreak=False,
                 randomization_seed=2,
             )
         )